src/plotter/requirements.txt
matplotlib==3.1.2
pyserial==3.4